Save and load StopwatchSeconds

+++ include/ClockModConfig.hpp
@@ -4,7 +4,7 @@
 DECLARE_CONFIG(ModConfig) {
 
     CONFIG_VALUE(ClockType,          int,    "Clock Type",                  0,     "Which time the Clock should display.");
-    CONFIG_VALUE(StopwatchSeconds,   double, "Stopwatch Seconds",           0,     "The saved time on the stopwatch.");
+    CONFIG_VALUE(StopwatchSeconds,   int,    "Stopwatch Seconds",           0,     "The saved time on the stopwatch.");
     CONFIG_VALUE(StopwatchPaused,    bool,   "Stopwatch Paused",            true,  "If the stopwatch is locked from incrementing.")
     CONFIG_VALUE(InSong,             bool,   "Show During Song",            true,  "If the Clock should be shown while playing a beatmap.");
     CONFIG_VALUE(InReplay,           bool,   "Show During Replay",          true,  "If the Clock should be shown while playing a replay.");
